+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ### Added
 
-- New menu for managing animals, access through the shop or through Better Game Menu context tabs.
+- New menu for managing animals, access through the shop or through Iconic Framework
 - More logging for broken textures.
 
 ### Fixed